# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cadastro de Livros API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/assunto/create": {
            "post": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Cadastra o Descricao do Assunto",
                "description": "Cadastra o Descricao do Assunto",
                "operationId": "77137bf163d4a6b0392de48af6d80ade",
                "parameters": [
                    {
                        "name": "Descricao",
                        "in": "query",
                        "description": "Insira o nome do assunto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/assunto/edit/{id}": {
            "get": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Retorna dados específicos de um Assunto",
                "description": "Retorna dados específicos de um Assunto",
                "operationId": "76b18cfa89a465f135a0e2d774c0795a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do assunto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/assunto/update/{id}": {
            "put": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Atualiza dados de um Assunto",
                "description": "Atualiza dados de um Assunto",
                "operationId": "7b846c629f49e00ef8aabc68ff6cfc71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do assunto",
                        "required": true
                    },
                    {
                        "name": "Descricao",
                        "in": "query",
                        "description": "Insira o nome do assunto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/assunto/delete/{id}": {
            "delete": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Deleta um Assunto específico",
                "description": "Deleta um Assunto específico",
                "operationId": "3f114aa21f46847c8133cb51905dc7da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do assunto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/autor/create": {
            "post": {
                "tags": [
                    "Autor"
                ],
                "summary": "Cadastra o Nome do Autor",
                "description": "Cadastra o Nome do Autor",
                "operationId": "a37ab1a10999fff1dff1e2f498b04bb2",
                "parameters": [
                    {
                        "name": "Nome",
                        "in": "query",
                        "description": "Insira o nome do autor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/autor/edit/{id}": {
            "get": {
                "tags": [
                    "Autor"
                ],
                "summary": "Retorna dados específicos de um Autor",
                "description": "Retorna dados específicos de um Autor",
                "operationId": "e5728089060e191c247a7413dda4fff2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do autor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/autor/update/{id}": {
            "put": {
                "tags": [
                    "Autor"
                ],
                "summary": "Atualiza dados de um Autor",
                "description": "Atualiza dados de um Autor",
                "operationId": "298f8700ca6da87131596df68e6764a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do autor",
                        "required": true
                    },
                    {
                        "name": "Nome",
                        "in": "query",
                        "description": "Insira o nome do autor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/autor/delete/{id}": {
            "delete": {
                "tags": [
                    "Autor"
                ],
                "summary": "Deleta um Autor específico",
                "description": "Deleta um Autor específico",
                "operationId": "25829d00157ae2a70c2039f3b788a349",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Insira o id do autor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}